msm: kgsl: Only clear enabled hfi interrupts
Currently, the act of clearing the interrupt bits
in the hfi handler can potentially clear an outstanding
hfi msgq interrupt, even though it is not part of the hfi
interrupt mask. This can race with the inline polling
of the hfi msgq interrupt and cause false timeouts.
Therefore, only clear the unmasked interrupts that are
expected to trigger the interrupt handler.

Change-Id: I51bb8f6eb2a9f04c23c60b5d0547b862bab926d5
Signed-off-by: Harshdeep Dhatt <hdhatt@codeaurora.org>
Signed-off-by: UtsavisGreat <utsavbalar1231@gmail.com>